requirements.txt
numpy==1.22.3
overcooked-ai==1.1.0
